add componentização do header, infelizmente não consegui o que queria com o breadcrumbs, mas pelo menos separei a logica para não ficar tão confuso
